+++ controllers/CourseController.php
@@ -7,6 +7,7 @@ use app\models\Course;
 use app\models\CourseSearch;
 use app\models\Helper;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\web\UploadedFile;
 use yii\filters\VerbFilter;
@@ -37,6 +38,10 @@ class CourseController extends Controller
      */
     public function actionIndex()
     {
+        if ( !\Yii::$app->user->can('course-list')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         $model = new Course(['scenario' => 'create']);
         if ($model->load(Yii::$app->request->post())) {
 
@@ -83,6 +88,10 @@ class CourseController extends Controller
      */
     public function actionView($id)
     {
+        if ( !\Yii::$app->user->can('course-view')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -93,8 +102,13 @@ class CourseController extends Controller
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
+/*
     public function actionCreate()
     {
+        if ( !\Yii::$app->user->can('course-create')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         $model = new Course();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -105,7 +119,7 @@ class CourseController extends Controller
             'model' => $model,
         ]);
     }
-
+*/
     /**
      * Updates an existing Course model.
      * If update is successful, the browser will be redirected to the 'view' page.
@@ -115,6 +129,10 @@ class CourseController extends Controller
      */
     public function actionUpdate($id)
     {
+        if ( !\Yii::$app->user->can('course-update')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post())) {
@@ -173,6 +191,10 @@ class CourseController extends Controller
      */
     public function actionDelete($id)
     {
+        if ( !\Yii::$app->user->can('course-delete')) {
+          throw new ForbiddenHttpException("Access denied");
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
